requirements.txt
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.3
cycler==0.11.0
execnet==1.9.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector==2.2.9
mysqlclient==2.1.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pyodbc==4.0.34
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.39
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
